Added config and auth. Using jwt. Request -> Fiter -> InternalFilter -> Validation -> ValidationService -> UserDetailsService / SecurityContectHoler.

+++ backend/src/main/java/com/shifterwebapp/shifter/course/CourseDto.java
@@ -26,7 +26,7 @@ public class CourseDto {
 
     private Float price;
 
-    private Float rating;
+    private Integer rating;
 
     private Integer ratingCount;
